requirements.txt
selenium==4.20.0
chromedriver-autoinstaller==0.6.4
beautifulsoup4==4.12.3
lxml==5.2.1
langchain==0.1.17
langchain-openai==0.1.5
python-dotenv==1.0.1